FIR: refactor base symbol computation for fake overrides
Not implemented override checker will use those to compute base symbols
for substitute/intersection overrides.
